Make image execute bash by default instead of rails
It turns out that there are more tools in this image that I would like
to use, other than just rails. One example is bundler. But sometimes I
just need to execute some ad-hoc commands in an interactive shell.

Therefore, I believe that the most convenient strategy is defining bash
as the default command to execute. For ONBUILD scenarios, rails server
is still a useful strategy.